v1.1.0 (iv) - Piglin prisoners now follow players when bribed with gold, only getting distracted by gold, repellents, and enemies - Removed amplified nether integration, as it does not exist in 1.16.5 TODO: - Integration with other 1.16.5 nether mods

+++ src/main/java/com/izofar/bygonenether/entity/ai/behvaior/ModAdmireItemTask.java
@@ -23,6 +23,6 @@ public class ModAdmireItemTask<E extends PiglinPrisonerEntity> extends Task<E> {
     }
 
     protected void start(ServerWorld pLevel, E pEntity, long pGameTime) {
-        pEntity.getBrain().setMemoryWithExpiry(MemoryModuleType.ADMIRING_ITEM, true, (long)this.admireDuration);
+        pEntity.getBrain().setMemoryWithExpiry(MemoryModuleType.ADMIRING_ITEM, true, this.admireDuration);
     }
 }
